Move assert_command outside of json (#367)
