+++ tests/commands/json/__init__.py
@@ -1,11 +1,10 @@
-from collections.abc import Callable, Sequence
+from functools import partial
 from typing import Any
-from unittest.mock import AsyncMock, Mock, call
+from unittest.mock import Mock
 
 from testfixtures import LogCapture
 
-from deebot_client.authentication import Authenticator
-from deebot_client.command import Command, CommandResult
+from deebot_client.command import CommandResult
 from deebot_client.commands.json.common import (
     ExecuteCommand,
     JsonSetCommand,
@@ -15,73 +14,12 @@ from deebot_client.event_bus import EventBus
 from deebot_client.events import EnableEvent, Event
 from deebot_client.hardware.deebot import FALLBACK, get_static_device_info
 from deebot_client.message import HandlingState
-from deebot_client.models import Credentials, DeviceInfo
+from tests.commands import assert_command as assert_command_base
 from tests.helpers import get_message_json, get_request_json, get_success_body
 
-
-def _wrap_command(command: Command) -> tuple[Command, Callable[[CommandResult], None]]:
-    result: CommandResult | None = None
-    execute_fn = command._execute
-
-    async def _execute(
-        _: Command,
-        authenticator: Authenticator,
-        device_info: DeviceInfo,
-        event_bus: EventBus,
-    ) -> CommandResult:
-        nonlocal result
-        result = await execute_fn(authenticator, device_info, event_bus)
-        return result
-
-    def verify_result(expected_result: CommandResult) -> None:
-        assert result == expected_result
-
-    command._execute = _execute.__get__(command)  # type: ignore[method-assign]
-    return (command, verify_result)
-
-
-async def assert_command(
-    command: Command,
-    json_api_response: dict[str, Any],
-    expected_events: Event | None | Sequence[Event],
-    command_result: CommandResult | None = None,
-) -> None:
-    command_result = command_result or CommandResult.success()
-    event_bus = Mock(spec_set=EventBus)
-    authenticator = Mock(spec_set=Authenticator)
-    authenticator.authenticate = AsyncMock(
-        return_value=Credentials("token", "user_id", 9999)
-    )
-    authenticator.post_authenticated = AsyncMock(return_value=json_api_response)
-    device_info = DeviceInfo(
-        {
-            "company": "company",
-            "did": "did",
-            "name": "name",
-            "nick": "nick",
-            "resource": "resource",
-            "deviceName": "device_name",
-            "status": 1,
-            "class": "get_class",
-        },
-        get_static_device_info(FALLBACK),
-    )
-
-    command, verify_result = _wrap_command(command)
-
-    await command.execute(authenticator, device_info, event_bus)
-
-    # verify
-    verify_result(command_result)
-    authenticator.post_authenticated.assert_called()
-    if expected_events:
-        if isinstance(expected_events, Sequence):
-            event_bus.notify.assert_has_calls([call(x) for x in expected_events])
-            assert event_bus.notify.call_count == len(expected_events)
-        else:
-            event_bus.notify.assert_called_once_with(expected_events)
-    else:
-        event_bus.notify.assert_not_called()
+assert_command = partial(
+    assert_command_base, static_device_info=get_static_device_info(FALLBACK)
+)
 
 
 async def assert_execute_command(
@@ -98,7 +36,9 @@ async def assert_execute_command(
     with LogCapture() as log:
         body = {"code": 500, "msg": "fail"}
         json = get_request_json(body)
-        await assert_command(command, json, None, CommandResult(HandlingState.FAILED))
+        await assert_command(
+            command, json, None, command_result=CommandResult(HandlingState.FAILED)
+        )
 
         log.check_present(
             (
