+++ tests/commands/json/test_clean_log.py
@@ -127,7 +127,7 @@ async def test_GetCleanLogs_analyse_logged(
         GetCleanLogs(),
         json,
         None,
-        CommandResult(HandlingState.ANALYSE_LOGGED),
+        command_result=CommandResult(HandlingState.ANALYSE_LOGGED),
     )
 
     assert (
@@ -142,7 +142,7 @@ async def test_GetCleanLogs_handle_error(caplog: pytest.LogCaptureFixture) -> No
         GetCleanLogs(),
         {},
         None,
-        CommandResult(HandlingState.ERROR),
+        command_result=CommandResult(HandlingState.ERROR),
     )
 
     assert (
