+++ tests/commands/json/test_custom.py
@@ -31,4 +31,4 @@ async def test_CustomCommand(
     expected: CustomCommandEvent | None,
     command_result: CommandResult,
 ) -> None:
-    await assert_command(command, json, expected, command_result)
+    await assert_command(command, json, expected, command_result=command_result)
